+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Print warning about unknown resource state before exiting when execution is interrupted with SIGINT.
 - Add `kubernetes nodegroup create`, `kubernetes nodegroup scale`, and `kubernetes nodegroup delete` commands (EXPERIMENTAL)
+- Added support for all shell completions provided by `cobra`.
+
+### Changed
+- Remove custom bash completion logic and replace it with `completion` command provided by `cobra`.
 
 ## [2.4.0] - 2022-12-19
 ### Added